+++ asset_allocation/risk_management/no_stop.py
@@ -17,8 +17,6 @@ class NoStop(stop_loss.StopLoss):
         frequency of return data
     rebelance: str
         rebalance frequency
-    rebalance_dates: list
-        list of rebalancing dates
     """
 
     def __init__(
@@ -27,12 +25,9 @@ class NoStop(stop_loss.StopLoss):
         stop_threshold: float,
         frequency: str,
         rebalance: str,
-        rebalance_dates: list,
         **kwargs,
     ) -> None:
-        super().__init__(
-            universe, stop_threshold, frequency, rebalance, rebalance_dates, **kwargs
-        )
+        super().__init__(universe, stop_threshold, frequency, rebalance, **kwargs)
 
     def assign(
         self,