+++ asset_allocation/strategies/pick_all.py
@@ -55,14 +55,9 @@ class PickAll(strategy.Strategy):
         self.stop_loss.assign(
             date=date, price_return=price_return, annualize_factor=annualize_factor
         )
-        # only calculate cov matrix on rebalance dates to save time
-        if date in self.rebalance_dates:
-            self.risk_engine.assign(
-                date=date, price_return=price_return, annualize_factor=annualize_factor
-            )
-            self.portfolio_risk_engine.assign(
-                date=date, price_return=price_return, annualize_factor=annualize_factor
-            )
+        self.risk_engine.assign(
+            date=date, price_return=price_return, annualize_factor=annualize_factor
+        )
 
     @property
     def selected_securities(self) -> np.ndarray: